Renamed Numbers module package structure
