+++ Numbers/src/main/kotlin/mathtools/numbers/IntegerMgr.kt
@@ -1,4 +1,4 @@
-package data
+package mathtools.numbers
 
 /** Data Structure for an integer that may travel within a fixed range
  * Developed by DK96-OS : 2019 - 2021 */
